requirements.txt
flask>=2.3.3
requests>=2.31.0
pandas>=2.1.0
tqdm>=4.66.1
matplotlib>=3.8.0
werkzeug>=2.3.7 
fastapi
python-dotenv>=1.0.0